pyproject.toml
[project]
name = "serverboizbot"
version = "0.1.0"
description = "Anything goes discord bot for ServerBoiz"
readme = "README.md"
requires-python = "~=3.11.0"
dependencies = [
    "discord-py[voice]>=2.6.4",
    "gitpython>=3.1.46",
    "pydantic>=2.12.5",
    "pypartpicker>=2.0.5",
    "pyrootutils>=1.0.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "dpytest>=0.7.0",
    "jinja2>=3.1.6",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 320

[tool.uv]
package = true

[project.scripts]
serverboizbot = "serverboizbot.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/serverboizbot"]

[tool.hatch.build.targets.sdist]
include = [
    "src/serverboizbot/assets/*.jpg",
    "src/serverboizbot/assets/*.gif",
    "src/serverboizbot/assets/*.webm",
    "src/serverboizbot/assets/*.png"
]
